+++ R/lmap.R
@@ -81,17 +81,14 @@ lmap <- function(.x, .f, ...) {
 
 #' @rdname lmap
 #' @export
-lmap_if <- function(.x, .p, .f, ...) {
-  sel <- which(probe(.x, .p))
-  lmap_at(.x, sel, .f, ...)
-}
-#' @rdname lmap
-#' @export
-lmap_if_else <- function(.x, .p, .if, .else, ...) {
+lmap_if <- function(.x, .p, .f, ..., .else = NULL) {
   sel <- probe(.x, .p)
 
-  .x <- lmap_at(.x, which(sel), .if, ...)
-  .x <- lmap_at(.x, which(!sel), .else, ...)
+  .x <- lmap_at(.x, which(sel), .f, ...)
+
+  if (!is_null(.else)) {
+    .x <- lmap_at(.x, which(!sel), .else, ...)
+  }
 
   .x
 }